feat: define service interfaces and repositories for user, book and reservation modules

+++ backend/src/main/java/com/hcmunre/library/entity/CuonSach.java
@@ -14,7 +14,7 @@ import lombok.*;
 public class CuonSach extends BaseEntity{
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer maCuonSach;
+    private Long maCuonSach;
 
     private String maVach;
     private String viTriKe;
